training/week-3/day-2/todo-management-app/requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
click==8.1.3
colorama==0.4.5
Flask==2.1.2
importlib-metadata==4.11.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg==3.0.15
psycopg-binary==3.0.15
psycopg-pool==3.1.1
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
tomli==2.0.1
tzdata==2022.1
Werkzeug==2.1.2
zipp==3.8.0
